Added bottom tab navigation. Moved homepage code to overviewpage
