+++ lib/views/settings.dart
@@ -3,31 +3,12 @@ import 'package:flutter/material.dart';
 class SettingsPage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: Colors.white,
-      appBar: AppBar(
-        elevation: 0.0,
-        centerTitle: true,
-        title: Text(
-          'Home',
-          style: TextStyle(
-            color: Colors.black,
-            fontFamily: 'Inconsolata',
-            fontSize: 25.0,
-          ),
-        ),
-        backgroundColor: Colors.white,
-        leading: IconButton(
-          onPressed: () {
-            Navigator.pop(context);
-          },
-          icon: Icon(Icons.arrow_back_ios),
-          color: Colors.black,
-        ),
-      ),
-      body: Column(
+    return Container(
+      child: Column(
         crossAxisAlignment: CrossAxisAlignment.center,
-        children: <Widget>[],
+        children: <Widget>[
+          Text('Settings')
+        ],
       ),
     );
   }
